Records follower count in the CSV file provided by the user
It uses logging package to print debug messages, instead of print.
It checks CSV file, provided by the user, for existence. If it exists it checks
its header and delimiter with the standard.
If the match fails it exits the script.
If the file does not exist it creates the file and writes the header on it.

Script then writes follower count in the CSV file with current time.